requirements.txt
# Dépendances principales
PyPDF2>=3.0.0
edsnlp>=0.7.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Dépendances de test
pytest>=7.0.0

# Utilitaires
python-dotenv==1.0.1
tqdm>=4.65.0
